Add files via upload
QuickSort is a widely used and efficient sorting algorithm that belongs to the category of divide-and-conquer algorithms. It was developed by Tony Hoare in 1960. The key idea of QuickSort is to choose a "pivot" element from the array and partition the other elements into two sub-arrays according to whether they are less than or greater than the pivot. The sub-arrays are then recursively sorted.

Here is sample code for quicksort hope this code help you to understand quick sort( I have placed comments in between the code for better understanding)